Podemos eliminar un usuario y ademas de eso tambien vamos a eliminar las contraseñas

+++ gestor-contrasenas-api/routes/password.routes.js
@@ -1,7 +1,7 @@
 const { Router } = require('express');
 const router = Router();
 
-const { postOne, getAll, putOne, deleteOne } = require('../controllers/password.controllers');
+const { postOne, getAll, putOne, deleteOne, deleteForUserId } = require('../controllers/password.controllers');
 const { validarJWT } = require('../middlewares/validar-jwt');
 
 router.use(validarJWT);
@@ -9,6 +9,8 @@ router.use(validarJWT);
 router.route('/')
     .post(postOne)
     .get(getAll)
+router.route('/userId')
+    .delete(deleteForUserId)
 router.route('/:id')
     .put(putOne)
     .delete(deleteOne)
